+++ lib/src/statement.dart
@@ -60,11 +60,11 @@ class ImportStmt extends Stmt {
   @override
   dynamic accept(StmtVisitor visitor) => visitor.visitImportStmt(this);
 
-  final String location;
+  final String path;
 
   final String nameSpace;
 
-  ImportStmt(this.location, {this.nameSpace});
+  ImportStmt(this.path, {this.nameSpace});
 }
 
 class VarDeclStmt extends Stmt {
@@ -73,6 +73,8 @@ class VarDeclStmt extends Stmt {
 
   @override
   dynamic accept(StmtVisitor visitor) => visitor.visitVarDeclStmt(this);
+
+  final String fileName;
 
   final Token name;
 
@@ -93,6 +95,7 @@ class VarDeclStmt extends Stmt {
   final bool isNamed;
 
   VarDeclStmt(
+    this.fileName,
     this.name, {
     this.declType = HT_Type.ANY,
     this.initializer,
@@ -192,7 +195,6 @@ class ContinueStmt extends Stmt {
 
 enum FuncStmtType {
   normal,
-  procedure,
   constructor,
   getter,
   setter,
@@ -206,6 +208,8 @@ class FuncDeclStmt extends Stmt {
 
   @override
   dynamic accept(StmtVisitor visitor) => visitor.visitFuncDeclStmt(this);
+
+  final String fileName;
 
   final Token keyword;
 
@@ -234,7 +238,7 @@ class FuncDeclStmt extends Stmt {
 
   final FuncStmtType funcType;
 
-  FuncDeclStmt(this.keyword, this.name, this.returnType, this.params,
+  FuncDeclStmt(this.fileName, this.keyword, this.name, this.returnType, this.params,
       {List<String> typeParams,
       this.arity = 0,
       this.definition,
@@ -269,6 +273,8 @@ class ClassDeclStmt extends Stmt {
   @override
   dynamic accept(StmtVisitor visitor) => visitor.visitClassDeclStmt(this);
 
+  final String fileName;
+
   final Token keyword;
 
   final String name;
@@ -285,8 +291,8 @@ class ClassDeclStmt extends Stmt {
 
   final List<FuncDeclStmt> methods;
 
-  ClassDeclStmt(this.keyword, this.name, this.superClass, this.superClassDeclStmt, this.superClassTypeArgs,
-      this.variables, this.methods,
+  ClassDeclStmt(this.fileName, this.keyword, this.name, this.superClass, this.superClassDeclStmt,
+      this.superClassTypeArgs, this.variables, this.methods,
       {List<String> typeParams}) {
     if (typeParams != null) this.typeParams.addAll(typeParams);
   }